Remove `refMatchSimilarities` calculation
This was used by the client to display violin plots, but this graph was removed ~2 months ago. For unknown reasons, this data was a major memory leak in the client (99% of a 1GB footprint of 1M reads). Removing this here doesn't identify the source of the leak, which is still a mystery to me, but it removes it by never giving this data to the client.