Allow identical filenames in different subdirectories
This allows, e.g. `barcode1/x.fastq` and `barcode2/x.fastq`, which is sometimes the output of guppy. This necessitated changes to how we track which files are seen, as before we were using the (extension-less) filename. We now incorporate the subdirectory (e.g. `barcode1/x`). As a consequence, annotated CSVs are stored in subdirectories corresponding to the FASTQ subdirectory structure.