Call snakemake with --cores argument
Snakemake accepts a --cores arg, and 5.11 made this compulsory so that our current snakemake calls will all fail if a user has snakemake 5.11 or higher.

Separately, PR 78 (https://github.com/artic-network/rampart/pull/78) shows an example of setting the number of threads for snakemake in the config file.

We now add `--cores` to each snakemake call, and set it to 1 _or_ the number of threads requested _if_ they are specified in the pipeline's snakemake config YAML or via the command line args for the snakemake call.